Fix - `signed_off_str` (#550)
* Adjusts offset of memory access in disassembler tests.

* Fixes minimal negative overflow in signed_off_str().

+++ sbpf/src/disassembler.rs
@@ -45,7 +45,7 @@ fn byteswap_str(name: &str, insn: &ebpf::Insn) -> String {
 #[inline]
 fn signed_off_str(value: i16) -> String {
     if value < 0 {
-        format!("-{:#x}", -value)
+        format!("-{:#x}", -(value as isize))
     } else {
         format!("+{value:#x}")
     }
